chore: rename files for publish
